change signature of getDistance returning float instead of unsigned int

+++ point.c
@@ -1,7 +1,7 @@
 #include "point.h"
 #include <math.h>
 
-unsigned int get_distance(Point pointA, Point pointB)
+float get_distance(Point pointA, Point pointB)
 {
   int diff_X = pointA.x - pointB.x;
   int diff_Y = pointA.y - pointB.y;
@@ -10,11 +10,11 @@ unsigned int get_distance(Point pointA, Point pointB)
 
 void get_closest_food(Point food_points[], int points_length, Point current_location, Point *closest_food_location)
 {
-  unsigned int closest_distance = get_distance(food_points[0], current_location);
+  float closest_distance = get_distance(food_points[0], current_location);
   *closest_food_location = food_points[0];
   FOR(1, points_length)
   {
-    unsigned int distance = get_distance(food_points[i], current_location);
+    float distance = get_distance(food_points[i], current_location);
     if (distance < closest_distance)
     {
       closest_distance = distance;
